Add oauth pages
